Update the embedder code
